+++ Editor/UploadOrchestrator.cs
@@ -245,6 +245,13 @@ namespace Anatawa12.ContinuousAvatarUploader.Editor
                     });
                     asset.Save();
                     WithTryCatch(() => OnUploadSingleAvatarFailed?.Invoke(asset, avatarToUpload, new List<Exception> { exception }));
+
+                    if (!asset.continueUploadOnError)
+                    {
+                        Log("Continue upload other avatars on error is disabled, so we are finishing the upload due to the error.");
+                        FinishUpload(asset, false);
+                        return;
+                    }
                 }
 
                 Log($"Avatar {asset.uploadingAvatarIndex + 1}/{asset.uploadSettings.Length} uploaded for platform {asset.uploadingTargetPlatform}.");