+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 ### Added
+- Allow user to stop building if any build or upload error occurred `#136`
 
 ### Changed
 - retry will not re-build avatar when the upload fails `#134`